# docs/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xenoteer Control API",
    "version": "1.0.0-phase4",
    "description": "Implemented Phase 4 HTTP control, observation, clipboard, capture, artifact, and viewer contract. Typed request and result definitions are generated from xenoteer-protocol and referenced from schemas/v1. Control WebSocket application messages are documented separately in websocket.md.",
    "license": {
      "name": "Apache License 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Default container-local control plane; production TLS may terminate at a trusted same-host proxy."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/desktops/{desktop_id}/clipboard/read": {
      "parameters": [{"$ref": "#/components/parameters/DesktopId"}],
      "post": {
        "operationId": "readClipboard",
        "tags": ["Clipboard"],
        "summary": "Read one bounded X11 selection representation",
        "description": "Requires clipboard:read. Text or small binary data may be inline; larger or policy-selected output is returned as a private artifact reference.",
        "parameters": [{"$ref": "#/components/parameters/DesktopGenerationQuery"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/clipboard-read-request.json"}}}
        },
        "responses": {
          "200": {
            "description": "Validated clipboard result. Cache-Control is private, no-store.",
            "content": {"application/json": {"schema": {"$ref": "../../../schemas/v1/clipboard-read-result.json"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "413": {"$ref": "#/components/responses/PayloadTooLarge"},
          "422": {"$ref": "#/components/responses/UnsupportedByTarget"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/CapabilityUnavailable"},
          "504": {"$ref": "#/components/responses/RequestOutcomeUnknown"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Exactly one long-lived API Bearer credential. The token is opaque and must not be placed in URLs. The closed grant vocabulary is desktop:status, desktop:observe, input:control, application:launch, application:terminate, window:control, clipboard:read, clipboard:write, capture:read, artifact:read, artifact:delete, and viewer:read. This scheme is not accepted by the viewer gateway WebSocket."
      },
      "viewerTicketSubprotocol": {
        "type": "apiKey",
        "in": "header",
        "name": "Sec-WebSocket-Protocol",
        "description": "A short-lived single-use viewer ticket carried only as the xenoteer.ticket.<secret> WebSocket subprotocol alongside the exact binary subprotocol. Never place the secret in a URL or log it."
      }
    },
    "parameters": {
      "DesktopId": {
        "name": "desktop_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "DesktopGenerationQuery": {
        "name": "desktop_generation",
        "in": "query",
        "required": true,
        "schema": {"type": "string", "format": "uuid"}
      },
...
    },
    "headers": {
...
      "RetryAfter": {
        "description": "Current implementation emits 1 second for nonterminal command results and backoff responses.",
        "required": true,
        "schema": {"type": "integer", "minimum": 1, "example": 1}
      },
      "OptionalRetryAfter": {
        "description": "Present with value 1 while the returned command result is nonterminal; absent when terminal.",
        "schema": {"type": "integer", "minimum": 1, "example": 1}
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Malformed JSON, invalid path/query/header, strict-shape failure, or unsupported protocol major.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Unauthorized": {
        "description": "A valid Bearer credential is required.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {"type": "string", "const": "Bearer realm=\"xenoteer\""}
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Forbidden": {
        "description": "The principal lacks the required grant, or a browser WebSocket Origin is denied.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "NotFound": {
        "description": "The requested versioned resource does not exist or may not be disclosed.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "Conflict": {
        "description": "Stale generation, lease conflict, command ID/content conflict, or unsafe cancellation stage.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/OptionalRetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"},
            "examples": {
              "staleReference": {"externalValue": "./examples/problem-stale-reference.json"}
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body exceeds the configured ordinary-message byte limit.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "UnsupportedByTarget": {
        "description": "The target cannot perform the structurally valid operation.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "TooManyRequests": {
        "description": "A bounded admission slot, queue, rate limit, or quota is full.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "CapabilityUnavailable": {
        "description": "The desktop or required subsystem is not ready.",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"}
        },
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"}
          }
        }
      },
      "RequestOutcomeUnknown": {
        "description": "The handler deadline elapsed after dispatch. This does not cancel accepted work.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "../../../schemas/v1/problem.json"},
            "examples": {
              "exactCommandRetry": {"externalValue": "./examples/problem-request-outcome-unknown.json"},
              "noAutomaticRetry": {"externalValue": "./examples/problem-request-outcome-unknown-no-retry.json"}
            }
          }
        }
      }
    }
  }
}